+++ public/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2017-09-27]
+### Fixed
+- MangaFox & MangaHere now use https URLs.
+
 ## [2017-09-23]
 ### Added
 - `[Backend]` Documentation for Base_Site_Model.